Remove legacy files and refactor codebase for consistency
- Deleted outdated and unused files: `.serena/project.yml`, `.vscode/launch.json`, `.github/workflows/codeql.yml`, `.serena/memories/porting/blockhelper_client_side_safety_2026-04-22.md`, and `.vscode/settings.json`.
- Removed the `LootInjector` class and associated logic for injecting custom loot tables.
- Refactored utility classes (e.g. `TargetHelper`, `EnchantHelper`) to leverage modern Java features like enhanced switch statements.
- Cleaned imports across the codebase for better readability and maintainability.
- Reduced redundancy and streamlined methods in classes like `ParticleHelper`, `EntityProjectileStatic`, and `BlockHelper`.
- Introduced `final` where applicable for constants, improving thread safety in entities and utilities.

+++ src/main/java/com/lulan/shincolle/utility/ParticleHelper.java
@@ -1,23 +1,8 @@
 package com.lulan.shincolle.utility;
 
-import com.lulan.shincolle.client.particle.ParticleChi;
-import com.lulan.shincolle.client.particle.ParticleCraning;
-import com.lulan.shincolle.client.particle.ParticleCube;
-import com.lulan.shincolle.client.particle.ParticleEmotion;
-import com.lulan.shincolle.client.particle.ParticleGradient;
-import com.lulan.shincolle.client.particle.ParticleLaser;
-import com.lulan.shincolle.client.particle.ParticleLightning;
-import com.lulan.shincolle.client.particle.ParticleLine;
-import com.lulan.shincolle.client.particle.ParticleSparkle;
-import com.lulan.shincolle.client.particle.ParticleSphereLight;
-import com.lulan.shincolle.client.particle.ParticleSpray;
-import com.lulan.shincolle.client.particle.ParticleStickyLightning;
-import com.lulan.shincolle.client.particle.ParticleSweep;
-import com.lulan.shincolle.client.particle.ParticleTeam;
-import com.lulan.shincolle.client.particle.ParticleTexts;
+import com.lulan.shincolle.client.particle.*;
 import com.lulan.shincolle.network.ModNetworking;
 import com.lulan.shincolle.network.S2CSpawnParticlePacket;
-
 import net.minecraft.client.Minecraft;
 import net.minecraft.client.multiplayer.ClientLevel;
 import net.minecraft.core.particles.ParticleTypes;
@@ -222,7 +207,7 @@ public class ParticleHelper {
             buf.writeDouble(y);
             buf.writeDouble(z);
             buf.writeDouble(0.3); // scale (lookX)
-            buf.writeDouble((double) teamId); // team indicator type (lookY)
+            buf.writeDouble(teamId); // team indicator type (lookY)
             buf.writeDouble(0.0); // unused (lookZ)
             byte[] payload = new byte[48];
             buf.getBytes(0, payload);
